backend/requirements.txt
# Core FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Database dependencies
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
PyJWT>=2.8.0

# AI and ML dependencies
openai>=1.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.0.20
langgraph-checkpoint-postgres>=2.0.0
langgraph-prebuilt>=0.6.0
llama-index>=0.9.0
pydantic-ai>=0.0.1
transformers>=4.36.0
torch>=2.1.0

# Arcade AI integration
langchain-arcade>=1.3.1

# Vector database
pgvector>=0.2.0

# File processing
python-docx>=1.1.0
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
aiofiles>=23.2.0

# HTTP client
httpx>=0.25.0
requests>=2.31.0

# Notification services
twilio>=8.10.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
python-dotenv>=1.0.0

# QR Code generation
qrcode[pil]>=7.4.0

# Voice capabilities for AI receptionist
speechrecognition>=3.10.0
pydub>=0.25.1
pygame>=2.5.0
openai-whisper>=20231117

# Vector embeddings and search
faiss-cpu>=1.7.4
langchain-openai>=0.1.0

# Web crawling and document processing
crawl4ai>=0.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
aiohttp>=3.9.0

# Enhanced document processing
python-docx>=1.1.0
PyPDF2>=3.0.0
python-pptx>=0.6.0
openpyxl>=3.1.0
pandas>=2.0.0

# LlamaIndex document processing
llama-parse>=0.4.0
llama-index>=0.10.0
llama-index-core>=0.10.0
llama-index-readers-file>=0.1.0

# Monitoring & Logging
structlog>=23.2.0
sentry-sdk[fastapi]>=1.38.0

# Rate Limiting
fastapi-limiter>=0.1.5
redis>=5.0.1

# Arcade MCP Integration
arcade-ai>=0.1.0
arcade-google>=0.1.0

# Email service dependencies
sendgrid>=6.10.0
email-validator>=2.1.0

